backend/requirements.txt
# Core Flask dependencies
Flask>=2.3.0
flask-cors>=4.0.0

# Database
pymongo>=4.5.0

# Authentication & Security
bcrypt>=4.0.0
PyJWT>=2.8.0

# Environment & Configuration
python-dotenv>=1.0.0

# File handling & Cloud storage
cloudinary>=1.36.0

# Machine Learning & Computer Vision
opencv-python>=4.8.0
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image/Video processing
imageio>=2.31.0
moviepy>=1.0.3

# Additional ML libraries (optional - TensorFlow may not be available for Python 3.13)
# tensorflow>=2.15.0

# Utilities
requests>=2.31.0